[FEATURE] add create Bill Impl

+++ src/main/java/com/enigma/wmbapi/dto/request/NewMenuRequest.java
@@ -15,5 +15,5 @@ public class NewMenuRequest {
     @NotNull
     String name;
     @Min(value = 0)
-    Float price;
+    Long price;
 }
